Feat[#250] : LaundryIndex
